% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_entry_check.R
\name{double_entry_check}
\alias{double_entry_check}
\title{Double data entry check}
\usage{
double_entry_check(x, y, id)
}
\arguments{
\item{x, y}{Two \code{data.frame}s with identical column names}

\item{id}{Character vector of column(s) to join the two \code{data.frame}s}
}
\value{
A \code{data.frame} with columns \code{id}, \code{field}, \code{x},
and \code{y}. The last two columns specify what the discrepancy is for each input
\code{data.frame}.
}
\description{
This function takes two data frames that have identical column names, 
which should also have idenctical records, and returns a table of errors where 
the two tables don't match up.
}
\examples{
repository_A <- data.frame(subject = rep(LETTERS[1:5], 2), visit = rep(1:5, 2), case = 1:10,
                           specimen = c(rep("DNA", 2), rep("Plasma", 3),
                           rep("Serum", 3), rep("RNA", 2)),
                           aliquots = 1:10, stringsAsFactors = FALSE)

repository_B <- data.frame(subject = c(rep(LETTERS[1:5], 2), rep(LETTERS[6], 2)),
                           visit = c(rep(1:5, 2), 6:7), case = 1:12,
                           specimen = c(rep("DNA", 2), rep("Plasma", 2), "RNA",
                                        rep("Serum", 2), rep("RNA", 2), rep("DNA", 3)),
                           aliquots = c(NA, 2:8, 12, 10, 11, 13), stringsAsFactors = FALSE)

errors <- double_entry_check(repository_A, repository_B, id = c("subject", "visit", "case"))
}

