\name{Reconcile-package}
\alias{Reconcile-package}
\alias{Reconcile}
\docType{package}
\title{
\packageTitle{Reconcile}
}
\description{
\packageDescription{Reconcile}
}
\details{

\packageDESCRIPTION{Reconcile}
\packageIndices{Reconcile}

}
\author{
\packageAuthor{Reconcile}

\packageMaintainer{Reconcile}
}

