% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcileSubjectVisits.R
\name{reconcileSubjectVisits}
\alias{reconcileSubjectVisits}
\title{Function for doing something/something comparisons}
\usage{
reconcileSubjectVisits(report.type = c("data.no.samples", "samples.no.data"),
  repository.data, site.data, repo.subject.column, repo.visit.column,
  site.subject.column, site.visit.column)
}
\arguments{
\item{report.type}{Which records to return}

\item{repository.data}{\code{data.frame} of repository data with columns that will provide unique rows of interest}

\item{site.data}{\code{data.frame} of site data with columns that will provide unique rows of interest}

\item{repo.subject.column}{Column name of subject IDs in repository \code{data.frame}}

\item{repo.visit.column}{Column name of visit in repository \code{data.frame}}

\item{site.subject.column}{Column name of subject IDs in site \code{data.frame}}

\item{site.visit.column}{Column name of visit in repository \code{data.frame}}
}
\value{
A \code{data.frame} with the appropriate column headings. If the \code{report.type} is \code{"data.no.samples"},
then the column headings are from the site \code{data.frame}. If the \code{report.type} is \code{"samples.no.data"},
then the column headings are from the repository \code{data.frame}
}
\description{
This function helps you see which subject/visit combinations are missing either in the site data or in the
biorepository data.
}
\examples{
repository <- data.frame(subject = c(LETTERS[1:5], LETTERS[1:5]), visit = c(1:5, 1:5),
                         specimen = c(rep("blood", 5), rep("DNA", 5)),
                         stringsAsFactors = FALSE)

site <- data.frame(patno = c(LETTERS[2:6], LETTERS[2:6]), clinevent = c(2:6, 2:6),
                   specimen = c("RNA", rep("blood", 4), "Serum", rep("DNA",  4)),
                   stringsAsFactors = FALSE)

# Check to see if there is data from the site but no samples at the biorepository
reconcileSubjectVisits(report.type = "data.no.samples", repository.data = repository,
                       site.data = site, repo.subject.column = "subject",
                       repo.visit.column = "visit", site.subject.column = "patno",
                       site.visit.column = "clinevent")

# Check to see if there are samples at the biorepository but no data from the site
reconcileSubjectVisits(report.type = "data.no.samples", repository.data = repository,
                       site.data = site, repo.subject.column = "subject",
                       repo.visit.column = "visit", site.subject.column = "patno",
                       site.visit.column = "clinevent")

}

