% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{reconcileVisits}
\alias{reconcileVisits}
\title{Checks if there are any visit data missing from site or biorepository for each subject.}
\usage{
reconcileVisits(repository.data, site.data, repo.subject.column,
  repo.visit.column, repo.specimen.column, site.subject.column,
  site.visit.column, site.specimen.column)
}
\arguments{
\item{repository.data}{\code{data.frame} of repository data with columns that will provide unique rows of interest}

\item{site.data}{\code{data.frame} of site data with columns that will provide unique rows of interest}

\item{repo.subject.column}{Column name of subject IDs in repository \code{data.frame}}

\item{repo.visit.column}{Column name of visit in repository \code{data.frame}}

\item{repo.specimen.column}{Column name of specimen in repository \code{data.frame}}

\item{site.subject.column}{Column name of subject IDs in site \code{data.frame}}

\item{site.visit.column}{Column name of visit in repository \code{data.frame}}

\item{site.specimen.column}{Column name of specimen in site \code{data.frame}}
}
\value{
A \code{data.frame} with columns \code{subject}, \code{visit}, \code{repository_specimen},
and \code{site_specimen}. An \code{NA} indicates that the corresponding specimen type was not found in
either the site data or biorepository data. For example, if a subject/visit contains \code{"DNA"} in the
\code{repository_specimen} column and \code{NA} in the \code{site_specimen} column, then the site data is
missing a DNA sample for that subject/visit. (This does not give any information about the number of samples
for each specimen--only whether or not at least a single specimen for each subject/visit is present in either data
set.)
}
\description{
This function looks at a subject/visit and helps you see if specimen data are missing from the
site or from the biorepository.
}
\examples{
repository <- data.frame(subject = c(LETTERS[1:5], LETTERS[1:5]), visit = c(1:5, 1:5),
                         specimen = c(rep("blood", 5), rep("DNA", 5)),
                         stringsAsFactors = FALSE)

site <- data.frame(patno = c(LETTERS[2:6], LETTERS[2:6]), clinevent = c(2:6, 2:6),
                   sample = c("RNA", rep("blood", 4), "Serum", rep("DNA",  4)),
                   stringsAsFactors = FALSE)

reconcileVisits(repository.data = repository, site.data = site, repo.subject.column = "subject",
                repo.visit.column = "visit", repo.specimen.column = "specimen", site.subject.column = "patno",
                site.visit.column = "clinevent", site.specimen.column = "sample")
}

